Make default `config`

+++ tests/test_run.py
@@ -4,14 +4,18 @@ import pathlib
 import shutil
 import subprocess
 
+import pytest
+
 from ikdsplit.utils import cd
 
 
-def test_run(tmp_path) -> None:
+@pytest.mark.parametrize("data_path", ["227", "002"])
+def test_run(data_path, tmp_path) -> None:
     """Test `run`."""
-    src = pathlib.Path(__file__).parent
+    src = pathlib.Path(__file__).parent / "testdata" / data_path
     fns = ["atoms_conventional.csv", "cell.dat", "ikdsplit.toml", "FPOSCAR"]
     for _ in fns:
-        shutil.copy2(src / _, tmp_path)
+        if (src / _).exists():
+            shutil.copy2(src / _, tmp_path)
     with cd(tmp_path):
         assert subprocess.call(["ikdsplit", "run", "-l", "2"]) == 0
